# api/openapi/recurring-mandates-service.yaml
openapi: 3.0.3
info:
  title: Recurring Mandates Service API
  version: 0.1.0
  description: >-
    API for variable recurring payment mandate lifecycle and execution
    (svc-pay-recurring-mandates). Tokens must carry the service id in aud; the
    TPP is the token's client (azp). DPoP (RFC 9449) is enforced: the
    Authorization header uses the DPoP scheme with a token bound by cnf.jkt,
    and every request carries a fresh proof (typ dpop+jwt, ES256 or PS256,
    htm, htu, iat within 5 minutes, ath, single-use jti). A plain Bearer
    token, a missing, replayed or mismatching proof is answered 401 with
    WWW-Authenticate DPoP error="invalid_dpop_proof".
servers:
  - url: https://api.example.com/open-finance/v1/vrp
    description: The service serves every path below under /open-finance/v1/vrp.
paths:
  /payment-consents:
    post:
      summary: Create recurring mandate consent
      operationId: createRecurringConsent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/JwsSignature'
        - $ref: '#/components/parameters/Dpop'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsentRequest'
      responses:
        '201':
          description: Consent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
  /payment-consents/{consentId}:
    get:
      summary: Get recurring mandate consent status
      operationId: getRecurringConsent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - name: consentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Consent status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
    delete:
      summary: Revoke recurring mandate consent
      operationId: revokeRecurringConsent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/Dpop'
        - name: consentId
          in: path
          required: true
          schema:
            type: string
        - name: reason
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Mandate revoked (also returned when it was already revoked)
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /payments:
    post:
      summary: Execute recurring payment under mandate
      operationId: executeRecurringPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/JwsSignature'
        - $ref: '#/components/parameters/Dpop'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutePaymentRequest'
      responses:
        '201':
          description: Payment accepted (or the original payment on an idempotent replay)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /payments/{paymentId}:
    get:
      summary: Get recurring payment status
      operationId: getRecurringPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
components:
  responses:
    Error:
      description: Error with a stable code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    InteractionId:
      name: x-fapi-interaction-id
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      required: true
      schema:
        type: string
    JwsSignature:
      name: x-jws-signature
      in: header
      required: true
      schema:
        type: string
    Dpop:
      name: DPoP
      in: header
      required: true
      description: DPoP proof JWT for this request, verified by the service (single use).
      schema:
        type: string
  schemas:
    CreateConsentRequest:
      type: object
      required: [Data]
      properties:
        Data:
          type: object
          required: [MaximumIndividualAmount, PeriodType, ExpirationDateTime]
          properties:
            MaximumIndividualAmount:
              type: string
            Currency:
              type: string
            PeriodType:
              type: string
            ExpirationDateTime:
              type: string
              format: date-time
            DebtorAccount:
              type: object
              description: >-
                Optional PSU account the collections debit. When present it is
                checked with the accounts service (active, debits allowed, same
                currency) at creation and before every collection.
              required: [Identification]
              properties:
                Identification:
                  type: string
    ExecutePaymentRequest:
      type: object
      required: [Data]
      properties:
        Data:
          type: object
          required: [ConsentId, InstructedAmount]
          properties:
            ConsentId:
              type: string
            InstructedAmount:
              type: object
              required: [Amount, Currency]
              properties:
                Amount:
                  type: string
                Currency:
                  type: string
    ConsentResponse:
      type: object
      properties:
        Data:
          type: object
          properties:
            ConsentId:
              type: string
            Status:
              type: string
    PaymentResponse:
      type: object
      properties:
        Data:
          type: object
          properties:
            PaymentId:
              type: string
            Status:
              type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: FORBIDDEN, NOT_FOUND, CONFLICT, CONCURRENT_UPDATE, BUSINESS_RULE_VIOLATION, INVALID_REQUEST, DEPENDENCY_UNAVAILABLE or INTERNAL_ERROR
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time